requirements.txt
streamlit
pandas
requests
beautifulsoup4
google-generativeai
openpyxl